+++ src/transaction.rs
@@ -5,14 +5,13 @@ use nom::{
     bytes::complete::{tag, take_while},
     character::complete::{char, space0, space1},
     combinator::{cut, iterator, map, opt, success, value},
-    combinator::{eof, not},
     sequence::{delimited, preceded, separated_pair, terminated, tuple},
     Parser,
 };
 
 use crate::{
-    account, account::Account, amount, amount::Amount, date, end_of_line, metadata, string, Date,
-    Decimal, IResult, Span,
+    account, account::Account, amount, amount::Amount, date, empty_line, end_of_line, metadata,
+    string, Date, Decimal, IResult, Span,
 };
 
 /// A transaction
@@ -175,13 +174,7 @@ fn do_parse<D: Decimal>(
         let (input, (tags, links)) = tags_and_links(input)?;
         let (input, _) = end_of_line(input)?;
         let (input, metadata) = metadata::parse(input)?;
-        let mut iter = iterator(
-            input,
-            alt((
-                posting.map(Some),
-                preceded(not(eof), end_of_line.map(|_| None)),
-            )),
-        );
+        let mut iter = iterator(input, alt((posting.map(Some), empty_line.map(|_| None))));
         let postings = iter.flatten().collect();
         let (input, _) = iter.finish()?;