+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed 
 
-* Comment or empty line between posting creating errors (#29) 
+* Error when there is a comment or empty line between postings (#29) 
+* Error when there is a comment or empty line between metadata 
 
 
 ## [1.0.0-alpha.11] - 2023-06-15